requirements.txt
fastapi
uvicorn
pydantic
google-generative-ai
python-dotenv
